git: var
