--- completers/git_completer/cmd/var_generated.go
@@ -1,17 +1,0 @@
-package cmd
-
-import (
-	"github.com/spf13/cobra"
-)
-
-var varCmd = &cobra.Command{
-	Use:   "var",
-	Short: "Show a Git logical variable",
-	Run: func(cmd *cobra.Command, args []string) {
-	},
-}
-
-func init() {
-
-	rootCmd.AddCommand(varCmd)
-}
